+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### BUG FIXES
 
+- `[blocksync]` tolerate late BlockResponse from honest peers after switching to consensus
+  ([\#5959](https://github.com/cometbft/cometbft/pull/5959))
 - `[consensus]` Fix `double_sign_check_height = 1` performing no double-sign
   checks due to off-by-one error in loop condition (`i < N` should be
   `i <= N`). The value `1` now correctly checks the previous block as intended.